Fix package initialization race-condition
Originally providers map was initialized in `init` function. It was
not initialized when GitHub provider was trying to add itself in its
own `init`. This is due the fact, that in Go, `init` functions within
package are most likely to be called in lexical filename order. This
is only encouraged by the spec, so cannot really rely on this too much.
Turned it into a explicit dependency to make sure, it's initialized
when first package tries to set it up.